refactor(tasks): decompose runner.ts into phase modules
